connector refactor for ArchUnit
