+++ digiwf-connector/digiwf-camunda-connector-service/src/test/java/de/muenchen/oss/digiwf/connector/ArchitectureTest.java
@@ -1,0 +1,12 @@
+package de.muenchen.oss.digiwf.connector;
+
+import com.tngtech.archunit.junit.AnalyzeClasses;
+import com.tngtech.archunit.junit.ArchTest;
+import com.tngtech.archunit.junit.ArchTests;
+import de.muenchen.oss.digiwf.archunit.HexagonalArchitectureTest;
+
+@AnalyzeClasses
+public class ArchitectureTest {
+    @ArchTest
+    static final ArchTests hexagonalArchitecture = ArchTests.in(HexagonalArchitectureTest.class);
+}
