+++ digiwf-connector/digiwf-camunda-rest-connector-starter/src/main/java/de/muenchen/oss/digiwf/connector/camunda/rest/DigiWFCamundaConnectorProperties.java
@@ -1,0 +1,19 @@
+package de.muenchen.oss.digiwf.connector.camunda.rest;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.validation.annotation.Validated;
+
+import java.util.List;
+
+@Getter
+@Setter
+@Validated
+@ConfigurationProperties(prefix = "de.muenchen.oss.digiwf.connector.adapter.camunda.rest")
+public class DigiWFCamundaConnectorProperties {
+
+    private List<String> filterVariables;
+
+
+}
