+++ digiwf-connector/digiwf-camunda-rest-connector-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-de.muenchen.oss.digiwf.connector.adapter.camunda.rest.DigiWFCamundaConnectorAutoConfiguration
+de.muenchen.oss.digiwf.connector.camunda.rest.DigiWFCamundaConnectorAutoConfiguration
