+++ digiwf-connector/digiwf-camunda-rest-connector-starter/src/test/java/de/muenchen/oss/digiwf/connector/adapter/in/camunda/rest/CamundaClientConfigurationTest.java
@@ -1,0 +1,33 @@
+package de.muenchen.oss.digiwf.connector.adapter.in.camunda.rest;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+
+class CamundaClientConfigurationTest {
+
+    private CamundaClientConfiguration config;
+    private List<String> initialFilteredVariables;
+
+    @BeforeEach
+    void setUp() {
+        initialFilteredVariables = Arrays.asList("var1", "var2", "var3");
+        config = new CamundaClientConfiguration(initialFilteredVariables);
+    }
+
+    @Test
+    void testGetFilters() {
+        List<String> expectedFilters = Arrays.asList("var1", "var2", "var3",
+                CamundaClientConfiguration.TOPIC_NAME,
+                CamundaClientConfiguration.TYPE_NAME,
+                CamundaClientConfiguration.INTEGRATION_NAME);
+        List<String> actualFilters = config.getFilters();
+
+        assertThat(actualFilters).isEqualTo(expectedFilters);
+    }
+}
